seperating out helpers
